pass chess move msgs between players
